[IMP] Got MARC XML stuff working smoothly
Fixed regexp's in parseIndicoID()
I was using search instead of match, also needed a $ on the end.
It was always matching conference regardless of whether it was conf, contrib or subcontrib

In createCDSRecord(), fixed args calling XXXXToXMLMarc21() methods

Tested on several events, contributions and subcontributions successfully

Added method getAllowedToAccessList() to class SubContribution.
It simply returns an empty list. This makes it work with getRecursiveAllowedToAccessList()