Let request definitions override notification tpls
It was already possible in a plugin (by simply providing a template with
the same name), but now it can also be done cleanly from the definition
which is necessary e.g. if a plugin contains more than one request
definition or if a request definition in the core wants to use its own
template.