Fix ordering of sessions in timetable (agenda)
The sessions are ordered first by start date and
secondly by end date (but reversed), to achive having
longer sessions on top of the list. If this ordering
is not applied, then a session that starts the same
hour as a session block that belongs to the first,
will be displayed *after* the session block, which
looks ugly.
Normally the session blocks should be displayed
differently than the parent sessions in the agenda
(indented).