Add new tracking sync endpoint
Unlike the previous sync functionality which submitted a form which
updated instance-related data and then synchronized the instance.
This is a dedicated endpoint, accessible through a simple GET request.
It does not require a form or data to be passed and will only sync the
existing data to the instance tracking server (cephalopod).