Simplify _checkSessionUser
Instead of setting a flag to stop processing we noy simply raise a
Forbidden exception which triggers a redirect to the login page in case
the user is not logged in.

Also show better errors when access to room booking has been denied
(module not enabled or user has no access) instead of the old generic
AccessError.