Set prefixes on protection forms
Those forms contain widgets that rely on unique IDs and since we open
the non-inheriting children dialog while a protection dialog is open we
may end up with two protection froms being visible at the same time.

In this specific case we know that we never have the same two protection
forms open so the IDs are unique enough like this.