Disable qTip animation of timetable balloons
Since elements in HTML have a non-configurable north-west gravity, they
always extend/shrink to/from the right and bottom. In our case since the
qTips are displayed on top of the timetable entries they thus need to be
moved every time the content changes size. This cause the animation of
the qTip to kick in and produce the weird up movement.