URL normalizations: fix multi-locator behavior
e.g. if there's an event_id, a track_id and an abstract_id.
both track and abstract are linked to an event and if they point
to different events it's impossible to fix the URL since we cannot
know which one to use, so we simply fail instead of letting such a case
pass through.