Add a GIN index on VCRoom data columns
That way json queries like VCRoom.data.contains({'vidyo_id': '12345'})
don't require a seq scan on the whole table.  Note that
VCRoom.data['vidyo_id'].astext does not use the index since it used the
`->>` operator which is not covered by it. Using `contains()` uses `@>`
which is covered and as long as you don't have to deal with different
data types it does the job just fine.