Introduce a parameterized Badge placeholder
To allow arbitrary fields from the registraion on badges/tickets etc,
introduce a parameterized placeholder for the designer.

Currently in this simplest form the parameter should be the
numeric field id, possibly followed by the subkey
separated with ':'.

To enable easier identification of the field id, it is now displayed
in the regform edit mode (I'll still need to add some styling there).

This is related to  issue #3146.