Improved TeX contribution list export
Handle different sort options and allow export of the generated tex
directory as zip-file as well.

For the zip export, run the normal processing and just zip up the
complete resulting folder.
We never abort in this case, even if the tex run fails as this may
help in debugging the tex problems.