Only allow access to regforms in the ACL
Otherwise there is a weird edge case if one regform is in the ACL and
another one isn't. In that case you'd get full access to both regforms
as soon as the one in the ACL got scheduled. And the other regform may
even let you proceed to payment, except that payment is not whitelisted
for restricted access (since unpaid registrants would already have event
access if that regform was in the ACL).

So instead we restrict the list of available regforms if you lack event
access to those that are actually in the ACL.