Do not set metadata when claiming files
It is rather meaningless in some cases, and we have relationships (or
rather their backrefs) when the File gets associated with another
object, so using the metadata to store additional information regarding
the claim is a bit pointless.

Besides this, the current metadata structure did not really take into
account a file being claimed by more than one object, and existing
metadata would have been overwritten.