Remove json_or_form_or_query meta location again
This doesn't work because we'd have to split any list in the JSON data
into MultiDict values. The previous logic also merged values from
different locations together which was messy.

This way we need to be explicit with the location a value comes from
when using webargs, but avoid all this ugly complexity.