Do not cache JS i18n data for invalid locales
Otherwise each request for an invalid locale will create a cache file.
Security scanners and similar tools may request a lot of invalid locales,
and thus create a large amount of small but useless files in the cache
folder (which are eventually cleaned up but not creating them to begin
with is clearly better).