Fix location tracking involving predefined rooms
The events on the FK columns only trigger during a flush, so just
wrapping the update of the objects in the tracking contextmanager wasn't
enough.

Now the tracking isn't done inside the operation functions but rather in
the RHs that may make changes to the location; similar to how it's
already the case for time change tracking.