Show placeholders in room-specific notification settings
Added notification placeholders in  the `Notifications` tab
of the `Edit Room Details` modal. The notification settings
were previously only available from the `api/admin/settings`
endpoint which is admin-protected.

This commit adds a new unprotected endpoint `api/config/notifications`
which exposes only the notification settings. This endpoint is
used to show the corresponding placeholder values in the input
elements.