Add `avatar_url` to user search results
The `avatar_url` field is added to results from `/user/search`
and from `/api/principals`. The modification of the second
endpoint is needed because the results are used by the user search
modal.

The `avatar_url` always contains a valid url. If the user
does not have a profile picture set, a default avatar url is
returned instead.