Show avatars in user search
If the user has set a profile picture, it will
be displayed in the search result instead of
the default result icon.

This change only affects user search even though
the code is shared between users and groups. Groups
still use the default `resultIcon` passed to `SearchFactory`,
if the `avatarURL` field of the search result does not exist.