Auto-expand notes from meeting note permalinks
Currently, the meeting notes permalinks lead
to a plain HTML page with just the notes.

This commit changes the permalink to point to
the contribution/block title fragment on the
event page (same link when clicking the direct
link next to the title). These new permalinks are
also used in search.

In addition, it also adds `?note=note_id` to the query
string. This query field is used to automatically
expand the relevant note when the event page is loaded.