Fix `ICSCalendarLink` button overflow (#4945)
* Show rounded corners for the last button only

* Make `ICSCalendarLink` scale with its buttons

Thanks to @javfg. Changes are based on this SO thread:
https://stackoverflow.com/q/55040250

In summary, the width of the popup is determined by the
width of the buttons. This is needed because the button
widths depend on the specific language.

The scaling behavior is achieved by setting `width: 0` and
`min-width: 100%` on the popup content, which lets the
buttons determine the width.