Make it less likely to accidentally create mixed categories
Currently, the category field in the event creation form
gets prefilled with the current category. This can lead
to creating events in categories which should only contain
other subcategories, if one is not careful. This could be
especially problematic in the case of root category.

This commit changes the behavior such that it does not
prefill categories which only contain other categories
and no events. In the case of non-root categories,
we also require that the user has create access
to the category in order to stop the prefilling.

This is to maintain the user friendliness of the
category picker for non-admin users as the picker
starts at the root category if no category is prefilled.