Timetable: add warning in case of draft mode
Adds a warning to the manage/timetable page of
conferences in case the draft mode is enabled.
This is the same warning shown on the main settings
page.

Instead of showing the 'Manage contributions' button,
we change the warning to include a 'Publish now', which
turns off the draft mode. We show the same modal that
is shown on the contribution settings page
('manage/contributions') when the button is clicked.