Add `reviewed_dt` to editing revisions (#4993)
* Add 'reviewed_dt' to EditingRevision table

For revisisons with a final state, the value
of 'reviewed_dt' is set to `created_dt`, otherwise
it is NULL.

When a new revision is made, `reviewed_dt` is updated
with the current time. When a review is deleted, the
value is set to NULL.

* Show 'reviewed_dt' for paper reviews.

We currently show 'created_dt' of the revision, which
is confusing if the review was made on a different date.

This change makes sure the correct time is displayed
for the review.