Force login before showing regform list if login required
In case a user is registered in an event that's restricted to registered
participants and accesses a deep link to an event page (e.g. timetable)
while not logged in, we now force them to login befor redirecting them
to the registration form(s). Like this the user will end up on the
intended page after logging in if they are already registered.