Fix duplicate imports of the moment library
For a reason yet unknown the webpack ProvidePlugin
does not import a library as a singleton but sometimes
duplicates it several times. This causes problems when
we need to set global properties on the moment object
such as moment.locale().

This commit removes the ProvidePlugin in favour of importing
the library manually and making it globally available via
window.moment.