Disable shared runtime chunk for plugins
Otherwise we get very weird errors that only happen if both indico and
plugins are built in production mode. It's not clear what exactly is
causing this problem, but this fixes it - apparently without any
negative side effects - so let's hope we can just keep it like this and
it doesn't introduce any new weird bugs.