Update RHEditingUploadContributionFile query params
Due to the nature of FileManager, the attributes that identify a file
depend on a `id` field which causes conflicts if we rely on two
different models (`File` and `PaperFile`).

For that reason, considering papers are an exceptional case, I renamed
the attribute within this domain in order for the upload request to
succeed.