Prevent calling editor.getData on every change
Calls to .getData() become too time consuming for
long texts and lag the editor as you type.

A (not so nice) workaround is to pass around a getter
for the data instead of the data itself. This eliminates
the .getData() calls but also breaks 'disabledUntilChange'
because we only access the data when submitting.