Clarify rejected message in editing timeline
When an author rejects proposed changes by an editor and submits
a new version themselves, the title of the revision item is still
'Submitter rejected proposed changes' which is misleading.

This changes the title to
'Submitter rejected proposed changes and uploaded a new revision'
when a new revision has been uploaded by the author.