Fix some API calls with personal tokens
When accessing APIs returning speaker information determining the sort
order for speakers accesses `session.user` (as `display_full_name`
uses a user preference), but doing so from an API call results in a call
to `get_request_user` which does not know about the API and thus does
another oauth scope check for the `everything` scopes, which fails if
the token used to authenticate the API call doesn't have that scope.

The workaround (considering that the API is legacy) is to always return
"no user" since legacy API calls are not really considered in the context
of a user (except for access checks) anyway.