Add watchfiles-based reloader
Watchman works great, but there's no proper release for Python 3.10+ and
it requires installing the watchman binary.

Watchfiles on the other hand is much easier to install as it includes
its (Rust-based) inotify backend in its wheel.

The only real drawback is that there's no magic that detects and fully
groups Git changes (e.g. checkouts). But debouncing covers this just
fine, and if needed the thresholds could be tuned (if some poor soul
develops on a super slow machine where Git checkouts take ages).