Add stylelint-stylistic rules (#5984)
Stylelint deprecated those rules in the core and recommends using
prettier instead. We already use stylelint-prettier, but the ancient
prettier version is no longer compatible with all our SCSS files
(`@forward` seems to break it), so we cannot rely on this anymore.

Upgrading prettier is not an option for now either because of how much
it screws up formatting in our JS code.

By using these rules we at least get some basic style enforcement, in
particular enforcing single quotes.