Fix forced-event-locale fallback
Forcing the event locale without taking the session into account is done
when sending emails to others, since those should pretty much never use
the user's current session locale but rather the event (or server)
default locale.

There's still an issue where registration notifications to an
unauthenticated user will not use their locale but the default one, but
this cannot be fixed without storing the locale used during registration
since at the time where we send a notification we do not know if the
user self-registered or it was a manager registering them, so we do not
know if it's safe to take the session locale into account.