Fix legacy styles leaking into roombooking CSS
This was caused by this unholy mix of mixins and actual CSS rules for
elements in _typography.scss, and the recently added import of that
module in the PlaceholderInfo stylesheet resulted in those rules from
being used in multiple modules, thus being added to common.css by
webpack.

Now it's properly split into two modules, one that can be safely
imported anywhere for its mixins, and one that actually defines all the
rules.