Fix error when accessing via link-local IP
Apparently this can happen when using localhost on certain
configurations where the IP then contains an interface specifier.
Without it the IP is not strictly unique, but nothing in indico expects
such identifiers, and it breaks when the IP goes into a postgres IP
column e.g. in case of the user's last login IP.