Fix edge-case with new and deleted revisions
This fixes an edge case where the conflict detection was not correctly triggered during the simultaneous edit of a new note, as well as fixing a case where attempting to create a new note revision after a deletion would raise an error. Unit tests were also re-written to validate this.