Fix splitting bookings
A previous PR introduced an incorrect change that reset the end date of
the old booking to the old end date and new end time, instead of that of
the last occurrence that already happened.

This caused 'reservation has no occurrences' errors because it made it
very likely for the new reservation created during the split to conflict
with the modified old reservation.