Fix __editable__ installs in devserver reloader (#6322)
When installing a plugin in editable mode but without using setuptools
(which happens when setuptools is not installed, when using `uv`, and
possibly other cases as well) there's no longer an egg-link inside
`site-packages` (which apparently gets added to `sys.path` implicitly),
but instead there's a `__editable__.*.pth` file that imports a special
module which contains a mapping between package names and filesystem
paths.

The werkzeug reloader does not know anything about this magic, and at
the time where we determine which paths to monitor plugins etc. have not
been imported yet so getting paths from `sys.modules` won't give us the
path for a plugin installed in such a way either.

We now explicitly grab a list of packages installed in editable mode,
and add their paths to the list of directories monitored by the reloader.