Fix error during direct DB upgrade from v2 to v3.3
We used the same `merge_map` temp table in two revisions, but since
there is no commit in between, it's still around when the second
revision runs. We now `DROP` it explicitly right after using it to
avoid this. Creating it `IF NOT EXISTS` would have been an alternative
options, but the benefit is minimal and alembic revisions should be
self-contained.