Add some missing checks for admin override mode
Admins were allowed to override nonbookable periods and bookable hours
regardless of whether they enabled admin override mode.

This is generally not a problem because the UI has checks as well, but
it's ugly and also made testing things harder (since just toggling off
admin override before creating the booking was not enough).

This fix only applies to creating bookings but not to modifying existing
bookings - in case of modifications there's some risk that if the
booking was created e.g. before the setup of NBPs, and then edited by
someone who does not have override permissions on the room.