Fix react using wrong translations in some cases
This happened when a Translate component got updated instead of fully
replaced, which happened e.g. in the ConflictIndicator when new data
was retrieved so it re-rendered with a different translated message.

In that case the same Translate component instance got reused by react,
just with different children. Due to a bug in react-jsx-i18n this did
not switch to the new source string to pick a translation, resulting
in the old translation still being used.

fixes #6469