Revert subscription change
Providing an empty subscription means the form should not re-render on
any state change, which is typically what we want. When no subscription
is specified, it's subscribed to everything, including `values`, which
means every single typed character in a text field in the form triggers
a re-render of the whole form (which results in extremely bad
performance).