Fix check for publishable files
The util was quite misleading since a revision was passed to it, but it
then got the latest revision with files from the editable, which at that
point was not yet flushed to the database. So doing a "request approval"
review on a revision that initially did not have publishable files but
adding them in that same revision was not possible.