Normalize empty strings and None in affiliations
Otherwise syncing affiliation data for a user may create duplicates,
because passing None in the constructor seems to be using the default
value, but then it gets stored as an empty string. So when querying it
again to avoid duplicates (filter by None), no result is returned,
and a new affiliation is inserted.