Hide participant list in conference meeting timetable
Instead, show the count and link to the participant list page.

Conference events are likely to have a large amount of participants
(easily hundreds), and inlining those on the meeting page not only
looks awful, but also results in hundreds of requests to show their
avatar images from everyone just accessing the timetable.