Preserve date when switching mode in RB DateNavigator
Currently, when changing e.g. from the day mode to a
week mode, the date changes as well. If the date is
in the middle of the week and we select the 'week' mode,
the date is changed to the start date of the week.

This means that for example selecting 'day' ->
'week' -> 'day' will not return to the same date which
is surprising.

This PR makes it so that the date does not change when
the view changes. This means that 'day' -> 'week' ->
'day' returns to the same date.