Fix scheduling existing contribs occasionally breaking (#6853)
This was a super obscure bug that only happened on the JACoW Indico
instance (don't ask me why), and only in production.

The legacy timetable code uses some UnderscoreJS functionality that is
not on par with lodash functionality, and for some reason (once again,
don't ask me why) in dev this page was using underscore, and in prod
this was using lodash.

Simply avoiding the chaining functionality and thus using only basic
functionality that works the same in both libraries fixes the issue.