Fix registering if required field is hidden
This was broken due to the schema-level validator not allowing empty
values for certain field types (e.g. text), but hidden fields sent
exactly that.

Now hidden fields are not sent at all, and on the server side we process
the schema twice, once without enforcing any required fields, and once
after knowing which fields are actually used based on conditions.

This commit also refactors the logic during registration updates, and
gets a list of hidden fields early on and deletes their data, instead of
setting it to an empty value.