Add flask-urls package to exports
When using `indico-url` imports this package is implicitly imported, and
in case of plugins we want to avoid it pulling in additional Indico
bundles as this can result in code modules being imported multiple
times.

For example, having a plugin with two webpack modules using `indico-url`
imports resulted in `custom_elements` code being executed twice, causing
a warning due to the (disallowed) duplicate registering of the same
custom element.