Use filter_choices when processing regform filters (#6946)
We had a hardcoded list which made it hard to properly filter by custom
fields coming from plugins. Now any field that returns filter choices
can be used for this propose.

This also fixes the country field not exposing its items for filtering,
and thus not allowing filtering even though it should have.

Co-authored-by: Adrian Moennich <adrian.moennich@cern.ch>