Sign user principal identifiers (#6952)
They always come from server-side data or from search results, and in
both cases they can be signed. By requiring a signature when loading
them, in particular in the endpoint to resolve them back to user
details, we ensure that it is not possible to look up users in bulk w/o
having already "seen" them.